[refactor] rename api endpoint.
